+++ src/main/java/com/cmc/suppin/fcm/domain/DeviceType.java
@@ -1,0 +1,5 @@
+package com.cmc.suppin.fcm.domain;
+
+public enum DeviceType {
+    ANDROID, IOS, OTHER
+}
